feat: cover unavailable menu case; add menu items fetch retries; use spinners; separate concerns
